Put metadata-related code into submodule
